+++ ExpAdd.java
@@ -6,7 +6,7 @@ public class ExpAdd extends ExpBinaryOperator
 		super("+", e1, e2);
     }
 
-    public int visit(Visitor v)
+    public Double visit(Visitor v)
 	{
 		return v.visitExpAdd(this);
     }
