+++ ExpSub.java
@@ -6,7 +6,7 @@ public class ExpSub extends ExpBinaryOperator
 		super("-", e1, e2);
     }
 
-    public int visit(Visitor v)
+    public Double visit(Visitor v)
 	{
 		return v.visitExpSub(this);
     }
